refactor(library/Attempt): extracts "Outcome/Failure.ts" from "Outcome/index.ts"
